requirement.txt
amqp==5.2.0
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
billiard==4.2.1
boto3==1.35.39
botocore==1.35.39
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
cloudscraper==1.2.71
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.1.2
django-appconf==1.0.6
django-autocomplete-light==3.11.0
django-celery-beat==2.7.0
django-ckeditor==6.7.2
django-ckeditor-5==0.2.17
django-cors-headers==4.4.0
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.6
django-extensions==3.2.3
django-js-asset==3.0.1
django-nine==0.2.7
django-redis==5.4.0
django-select2==8.2.1
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
elastic-transport==8.15.1
elasticsearch==8.15.1
elasticsearch-dsl==8.15.4
filelock==3.18.0
flower==2.0.1
fsspec==2025.3.0
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
gunicorn==23.0.0
h11==0.14.0
httplib2==0.22.0
huggingface-hub==0.29.3
humanize==4.11.0
idna==3.10
inflection==0.5.1
isodate==0.7.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
kombu==5.4.2
Levenshtein==0.23.0
lxml==5.3.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==2.1.2
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
pillow==10.4.0
platformdirs==4.3.6
prometheus_client==0.21.0
prompt_toolkit==3.0.48
proto-plus==1.24.0
protobuf==5.28.2
psycopg==3.2.3
psycopg2==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyJWT==2.9.0
pyparsing==3.1.4
PySocks==1.7.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.12.2
redis==5.1.1
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rsa==4.9
s3transfer==0.10.3
safetensors==0.5.3
scikit-learn==1.5.2
scipy==1.14.1
selenium==4.25.0
sentencepiece==0.2.0
sentry-sdk==2.16.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.1
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.1
torch==2.6.0
tornado==6.4.1
tqdm==4.67.1
transformers==4.49.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==2.2.3
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==4.0.2
zeep==4.3.1
websocket-client==1.8.0
whitenoise==6.8.2
wsproto==1.2.0
zeep==4.3.1
